+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Added `value` and `name` props to make checkbox act more like a real HTML5 checkbox. Previously
+name was generated from `id` prop.
 
 ## 1.4.2pm 
 * Fixed a styling issue that only occurred in Firefox